Added the Interner trait and split up interface.rs
